updated controllers folder

+++ src/main/java/org/yearup/controllers/CategoriesController.java
@@ -10,6 +10,7 @@ import org.yearup.data.ProductDao;
 import org.yearup.models.Category;
 import org.yearup.models.Product;
 
+import java.util.ArrayList;
 import java.util.List;
 
 // add the annotations to make this a REST controller
@@ -33,7 +34,7 @@ public class CategoriesController {
 
 
     // add the appropriate annotation for a get action
-    @GetMapping("{id}")
+    @GetMapping()
     @PreAuthorize("permitAll")
     public List<Category> getAll() {
         // find and return all categories
@@ -64,17 +65,33 @@ public class CategoriesController {
     @GetMapping("{categoryId}/products")
     @PreAuthorize("permitAll")
     public List<Product> getProductsById(@PathVariable int categoryId) {
-        // get a list of product by categoryId
-        return productDao.listByCategoryId(categoryId);
+        try {    // get a list of product by categoryId
+            var product = productDao.getById(categoryId);
+
+            if (product == null)
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+
+
+            return productDao.listByCategoryId(categoryId);
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Ooops... our bad.");
+        }
+
     }
 
     // add annotation to call this method for a POST action
     // add annotation to ensure that only an ADMIN can call this function
     @PostMapping()
     @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @ResponseStatus(value = HttpStatus.CREATED)
     public Category addCategory(@RequestBody Category category) {
         // insert the category
-        return categoryDao.create(category);
+        try {
+            return categoryDao.create(category);
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops ... our bad");
+        }
+
     }
 
     // add annotation to call this method for a PUT (update) action - the url path must include the categoryId
@@ -83,7 +100,12 @@ public class CategoriesController {
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     public void updateCategory(@PathVariable int id, @RequestBody Category category) {
         // update the category by id
-        categoryDao.update(id, category);
+     try {
+         categoryDao.update(id, category);
+     }
+     catch (Exception e){
+         throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+     }
 
     }
 
@@ -92,8 +114,17 @@ public class CategoriesController {
     // add annotation to ensure that only an ADMIN can call this function
     @DeleteMapping("{id}")
     @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @ResponseStatus(value = HttpStatus.NO_CONTENT)
     public void deleteCategory(@PathVariable int id) {
-        // delete the category by id
-        categoryDao.delete(id);
+        try {
+            var category = productDao.getById(id);
+
+            if (category == null)
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            categoryDao.delete(id);
+        }
+        catch (Exception e){
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
 }
